tls: splitting context code into separate files (#34061)
