+++ source/common/tls/context_impl.h
@@ -172,77 +172,6 @@ protected:
 
 using ContextImplSharedPtr = std::shared_ptr<ContextImpl>;
 
-class ClientContextImpl : public ContextImpl, public Envoy::Ssl::ClientContext {
-public:
-  ClientContextImpl(Stats::Scope& scope, const Envoy::Ssl::ClientContextConfig& config,
-                    Server::Configuration::CommonFactoryContext& factory_context);
-
-  absl::StatusOr<bssl::UniquePtr<SSL>>
-  newSsl(const Network::TransportSocketOptionsConstSharedPtr& options) override;
-
-private:
-  int newSessionKey(SSL_SESSION* session);
-
-  const std::string server_name_indication_;
-  const bool allow_renegotiation_;
-  const bool enforce_rsa_key_usage_;
-  const size_t max_session_keys_;
-  absl::Mutex session_keys_mu_;
-  std::deque<bssl::UniquePtr<SSL_SESSION>> session_keys_ ABSL_GUARDED_BY(session_keys_mu_);
-  bool session_keys_single_use_{false};
-};
-
-enum class OcspStapleAction { Staple, NoStaple, Fail, ClientNotCapable };
-
-class ServerContextImpl : public ContextImpl, public Envoy::Ssl::ServerContext {
-public:
-  ServerContextImpl(Stats::Scope& scope, const Envoy::Ssl::ServerContextConfig& config,
-                    const std::vector<std::string>& server_names,
-                    Server::Configuration::CommonFactoryContext& factory_context,
-                    Ssl::ContextAdditionalInitFunc additional_init);
-
-  // Select the TLS certificate context in SSL_CTX_set_select_certificate_cb() callback with
-  // ClientHello details. This is made public for use by custom TLS extensions who want to
-  // manually create and use this as a client hello callback.
-  enum ssl_select_cert_result_t selectTlsContext(const SSL_CLIENT_HELLO* ssl_client_hello);
-
-  // Finds the best matching context. The returned context will have the same lifetime as
-  // this ``ServerContextImpl``.
-  std::pair<const Ssl::TlsContext&, OcspStapleAction> findTlsContext(absl::string_view sni,
-                                                                     bool client_ecdsa_capable,
-                                                                     bool client_ocsp_capable,
-                                                                     bool* cert_matched_sni);
-
-private:
-  // Currently, at most one certificate of a given key type may be specified for each exact
-  // server name or wildcard domain name.
-  using PkeyTypesMap = absl::flat_hash_map<int, std::reference_wrapper<Ssl::TlsContext>>;
-  // Both exact server names and wildcard domains are part of the same map, in which wildcard
-  // domains are prefixed with "." (i.e. ".example.com" for "*.example.com") to differentiate
-  // between exact and wildcard entries.
-  using ServerNamesMap = absl::flat_hash_map<std::string, PkeyTypesMap>;
-
-  void populateServerNamesMap(Ssl::TlsContext& ctx, const int pkey_id);
-
-  using SessionContextID = std::array<uint8_t, SSL_MAX_SSL_SESSION_ID_LENGTH>;
-
-  int alpnSelectCallback(const unsigned char** out, unsigned char* outlen, const unsigned char* in,
-                         unsigned int inlen);
-  int sessionTicketProcess(SSL* ssl, uint8_t* key_name, uint8_t* iv, EVP_CIPHER_CTX* ctx,
-                           HMAC_CTX* hmac_ctx, int encrypt);
-  bool isClientEcdsaCapable(const SSL_CLIENT_HELLO* ssl_client_hello);
-  bool isClientOcspCapable(const SSL_CLIENT_HELLO* ssl_client_hello);
-  OcspStapleAction ocspStapleAction(const Ssl::TlsContext& ctx, bool client_ocsp_capable);
-
-  SessionContextID generateHashForSessionContextId(const std::vector<std::string>& server_names);
-
-  const std::vector<Envoy::Ssl::ServerContextConfig::SessionTicketKey> session_ticket_keys_;
-  const Ssl::ServerContextConfig::OcspStaplePolicy ocsp_staple_policy_;
-  ServerNamesMap server_names_map_;
-  bool has_rsa_{false};
-  bool full_scan_certs_on_sni_mismatch_;
-};
-
 } // namespace Tls
 } // namespace TransportSockets
 } // namespace Extensions
