+++ source/common/tls/server_context_impl.h
@@ -1,0 +1,95 @@
+#pragma once
+
+#include <openssl/safestack.h>
+
+#include <array>
+#include <deque>
+#include <functional>
+#include <memory>
+#include <string>
+#include <vector>
+
+#include "envoy/network/transport_socket.h"
+#include "envoy/ssl/context.h"
+#include "envoy/ssl/context_config.h"
+#include "envoy/ssl/private_key/private_key.h"
+#include "envoy/ssl/ssl_socket_extended_info.h"
+#include "envoy/stats/scope.h"
+#include "envoy/stats/stats_macros.h"
+
+#include "source/common/common/matchers.h"
+#include "source/common/stats/symbol_table.h"
+#include "source/common/tls/cert_validator/cert_validator.h"
+#include "source/common/tls/context_impl.h"
+#include "source/common/tls/context_manager_impl.h"
+#include "source/common/tls/ocsp/ocsp.h"
+#include "source/common/tls/stats.h"
+
+#include "absl/synchronization/mutex.h"
+#include "openssl/ssl.h"
+#include "openssl/x509v3.h"
+
+#ifdef ENVOY_ENABLE_QUIC
+#include "quiche/quic/core/crypto/proof_source.h"
+#endif
+
+namespace Envoy {
+namespace Extensions {
+namespace TransportSockets {
+namespace Tls {
+
+enum class OcspStapleAction { Staple, NoStaple, Fail, ClientNotCapable };
+
+class ServerContextImpl : public ContextImpl, public Envoy::Ssl::ServerContext {
+public:
+  ServerContextImpl(Stats::Scope& scope, const Envoy::Ssl::ServerContextConfig& config,
+                    const std::vector<std::string>& server_names,
+                    Server::Configuration::CommonFactoryContext& factory_context,
+                    Ssl::ContextAdditionalInitFunc additional_init);
+
+  // Select the TLS certificate context in SSL_CTX_set_select_certificate_cb() callback with
+  // ClientHello details. This is made public for use by custom TLS extensions who want to
+  // manually create and use this as a client hello callback.
+  enum ssl_select_cert_result_t selectTlsContext(const SSL_CLIENT_HELLO* ssl_client_hello);
+
+  // Finds the best matching context. The returned context will have the same lifetime as
+  // this ``ServerContextImpl``.
+  std::pair<const Ssl::TlsContext&, OcspStapleAction> findTlsContext(absl::string_view sni,
+                                                                     bool client_ecdsa_capable,
+                                                                     bool client_ocsp_capable,
+                                                                     bool* cert_matched_sni);
+
+private:
+  // Currently, at most one certificate of a given key type may be specified for each exact
+  // server name or wildcard domain name.
+  using PkeyTypesMap = absl::flat_hash_map<int, std::reference_wrapper<Ssl::TlsContext>>;
+  // Both exact server names and wildcard domains are part of the same map, in which wildcard
+  // domains are prefixed with "." (i.e. ".example.com" for "*.example.com") to differentiate
+  // between exact and wildcard entries.
+  using ServerNamesMap = absl::flat_hash_map<std::string, PkeyTypesMap>;
+
+  void populateServerNamesMap(Ssl::TlsContext& ctx, const int pkey_id);
+
+  using SessionContextID = std::array<uint8_t, SSL_MAX_SSL_SESSION_ID_LENGTH>;
+
+  int alpnSelectCallback(const unsigned char** out, unsigned char* outlen, const unsigned char* in,
+                         unsigned int inlen);
+  int sessionTicketProcess(SSL* ssl, uint8_t* key_name, uint8_t* iv, EVP_CIPHER_CTX* ctx,
+                           HMAC_CTX* hmac_ctx, int encrypt);
+  bool isClientEcdsaCapable(const SSL_CLIENT_HELLO* ssl_client_hello);
+  bool isClientOcspCapable(const SSL_CLIENT_HELLO* ssl_client_hello);
+  OcspStapleAction ocspStapleAction(const Ssl::TlsContext& ctx, bool client_ocsp_capable);
+
+  SessionContextID generateHashForSessionContextId(const std::vector<std::string>& server_names);
+
+  const std::vector<Envoy::Ssl::ServerContextConfig::SessionTicketKey> session_ticket_keys_;
+  const Ssl::ServerContextConfig::OcspStaplePolicy ocsp_staple_policy_;
+  ServerNamesMap server_names_map_;
+  bool has_rsa_{false};
+  bool full_scan_certs_on_sni_mismatch_;
+};
+
+} // namespace Tls
+} // namespace TransportSockets
+} // namespace Extensions
+} // namespace Envoy
