+++ test/common/quic/envoy_quic_proof_verifier_test.cc
@@ -3,6 +3,7 @@
 
 #include "source/common/network/transport_socket_options_impl.h"
 #include "source/common/quic/envoy_quic_proof_verifier.h"
+#include "source/common/tls/client_context_impl.h"
 #include "source/common/tls/context_config_impl.h"
 
 #include "test/common/config/dummy_config.pb.h"
