+++ tools/code_format/config.yaml
@@ -163,6 +163,7 @@ paths:
     - source/common/grpc/google_grpc_creds_impl.cc
     - source/common/local_reply/local_reply.cc
     - source/common/tls/context_impl.cc
+    - source/common/tls/server_context_impl.cc
     - source/common/tls/context_config_impl.cc
     - source/common/tls/ocsp/ocsp.cc
     - source/common/config/watched_directory.cc
